% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accession.R
\docType{class}
\name{Accession-class}
\alias{Accession-class}
\title{Accession Class}
\description{
An S4 Class to represent a breeDBase accession
}
\details{
The `accession_name` and `species_name` are required, all other fields are optional.
}
\section{Slots}{

\describe{
\item{\code{accession_name}}{Accession Name (must be unique)}

\item{\code{species_name}}{Species Name (must exist in the database)}

\item{\code{properties}}{(optional) a list of additional accession stock properties, where the list key is a stock property name supported by the breedbase server and the list value is either a string value or a vector of string values.}
}}

\seealso{
Other Accession: 
\code{\link{Accession}},
\code{\link{Pedigree-class}},
\code{\link{Pedigree}},
\code{\link{buildAccessionTemplate}()},
\code{\link{buildPedigreeTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writeAccessionTemplate}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
