% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessions.R
\name{buildAccessionTemplate}
\alias{buildAccessionTemplate}
\title{Build Accession Template}
\usage{
buildAccessionTemplate(accessions = NULL)
}
\arguments{
\item{accessions}{Vector of Accessions to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBase upload 
template for the provided accessions
}
\seealso{
Other Accession: 
\code{\link{Accession-class}},
\code{\link{Accession}},
\code{\link{Pedigree-class}},
\code{\link{Pedigree}},
\code{\link{buildPedigreeTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writeAccessionTemplate}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
