% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdatas.R
\name{buildPlotDataTemplate}
\alias{buildPlotDataTemplate}
\title{Build PlotData Template}
\usage{
buildPlotDataTemplate(plotdata = NULL)
}
\arguments{
\item{plotdata}{Vector of PlotData to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBasae upload
template for the provided plotdata.  This will be a 'Simple' 
spreadsheet format with data recorded at the plot-level.
}
