% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenies.R
\name{buildProgenyTemplate}
\alias{buildProgenyTemplate}
\title{Build Progeny Template}
\usage{
buildProgenyTemplate(progenies = NULL)
}
\arguments{
\item{progenies}{Vector of Progenies to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBase upload 
template for the provided progenies
}
