% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_buac_gm2}
\alias{convert_buac_gm2}
\title{Convert bu/ac to g/m2}
\usage{
convert_buac_gm2(x, crop = NULL)
}
\arguments{
\item{x}{grain yield in bu/ac}

\item{crop}{crop type ("barley", "oat", "wheat")}
}
\value{
grain yield in g/m2
}
\description{
Convert grain yield units from bu/ac to g/m2
}
