% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_buac_kgHa}
\alias{convert_buac_kgHa}
\title{Convert bu/ac to kg/Ha}
\usage{
convert_buac_kgHa(x, crop = NULL)
}
\arguments{
\item{x}{grain yield in bu/ac}

\item{crop}{crop type ("barley", "oat", "wheat")}
}
\value{
grain yield in kg/Ha
}
\description{
Convert grain yield units from bu/ac to kg/Ha
}
