% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_gm2_kgHa}
\alias{convert_gm2_kgHa}
\title{Convert g/m2 to kg/Ha}
\usage{
convert_gm2_kgHa(x)
}
\arguments{
\item{x}{grain yield in g/m2}
}
\value{
grain yield in kg/Ha
}
\description{
Convert grain yield units from g/m2 to kg/Ha
}
\seealso{
Other conversions: 
\code{\link{convert_buac_gm2}()},
\code{\link{convert_buac_kgHa}()},
\code{\link{convert_in_cm}()},
\code{\link{convert_kgHa_gm2}()},
\code{\link{convert_kghl_gL}()},
\code{\link{convert_lbsbu_gL}()},
\code{\link{convert_scale}()},
\code{\link{convert_sqft_m2}()}
}
\concept{conversions}
