% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_in_cm}
\alias{convert_in_cm}
\title{Convert in to cm}
\usage{
convert_in_cm(x)
}
\arguments{
\item{x}{height in inches}
}
\value{
height in cm
}
\description{
Convert height / length units from in to cm
}
\seealso{
Other conversions: 
\code{\link{convert_buac_gm2}()},
\code{\link{convert_buac_kgHa}()},
\code{\link{convert_gm2_kgHa}()},
\code{\link{convert_kgHa_gm2}()},
\code{\link{convert_kghl_gL}()},
\code{\link{convert_lbsbu_gL}()},
\code{\link{convert_scale}()},
\code{\link{convert_sqft_m2}()}
}
\concept{conversions}
