% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_kgHa_gm2}
\alias{convert_kgHa_gm2}
\title{Convert kg/Ha to g/m2}
\usage{
convert_kgHa_gm2(x)
}
\arguments{
\item{x}{grain yield in kg/Ha}
}
\value{
grain yield in g/m2
}
\description{
Convert grain yield units from kg/Ha to g/m2
}
\seealso{
Other conversions: 
\code{\link{convert_buac_gm2}()},
\code{\link{convert_buac_kgHa}()},
\code{\link{convert_gm2_kgHa}()},
\code{\link{convert_in_cm}()},
\code{\link{convert_kghl_gL}()},
\code{\link{convert_lbsbu_gL}()},
\code{\link{convert_scale}()},
\code{\link{convert_sqft_m2}()}
}
\concept{conversions}
