% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_lbsbu_gL}
\alias{convert_lbsbu_gL}
\title{Convert lbs/bu to g/L}
\usage{
convert_lbsbu_gL(x)
}
\arguments{
\item{x}{test weight in lbs/bu}
}
\value{
test weight in g/L
}
\description{
Convert test weight units from lbs/bu to g/L
}
