% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_scale}
\alias{convert_scale}
\title{Convert scale}
\usage{
convert_scale(x, oldmin, oldmax, newmin, newmax)
}
\arguments{
\item{x}{the original value}

\item{oldmin}{the old range minimum}

\item{oldmax}{the old range maximum}

\item{newmin}{the new range minimum}

\item{newmax}{the new range maximum}
}
\value{
the value in the new range
}
\description{
Convert a value from an ordinal scale in one range to a
value in a new range (0-->5, 1-->10)
}
