% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{createAccessionSearchDB}
\alias{createAccessionSearchDB}
\title{Create Accession Search Database}
\usage{
createAccessionSearchDB(name, address, version, call_limit = 10)
}
\arguments{
\item{name}{Database Name}

\item{address}{Database BrAPI address}

\item{version}{Database BrAPI version}

\item{call_limit}{(optional) The maximum number of concurrent BrAPI calls}
}
\value{
Accession Search Database, a list with the following names:
\describe{
  \item{name}{Database name}
  \item{address}{Database BrAPI URL}
  \item{version}{Database BrAPI version}
  \item{call_limit}{Max number of concurrent connections to the Databse}
}
}
\description{
Create a custom Accession Search Database with the specified 
database properties
}
\seealso{
Other accessionSearch: 
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
