% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{geocodeLocation}
\alias{geocodeLocation}
\title{Geocode Location}
\usage{
geocodeLocation(location)
}
\arguments{
\item{location}{Location to parse (ie Ithaca, NY; Caldwell Dr, Ithaca, NY; etc...)}
}
\value{
list(latitude, longitude, altitude)
}
\description{
Lookup the coordinates and elevation for the specified 
location.  The location text will be parsed by DataScienceToolkit's
`street2coordinates` API into coordinates and the elevation will be 
queried using the `coordinates2statistics` API.
}
