% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{getAccessionSearchCacheStatus}
\alias{getAccessionSearchCacheStatus}
\title{Get Accession Search Cache Status}
\usage{
getAccessionSearchCacheStatus(db)
}
\arguments{
\item{db}{Accession Search Database}
}
\value{
Cache status information, a list with the following names:
\describe{
  \item{saved}{timestamp of when the cache was last saved}
  \item{terms}{the number of database terms saved}
}
}
\description{
Get the status of the cached database terms for the provided Accession Search Database.
}
\details{
The cache status includes the timestamp of when the database terms were last loaded 
and the total number of database terms saved.  Use the \code{\link{updateAccessionSearchCache}} 
function to update the cached database terms for an Accession Search Database.
}
\seealso{
\link{updateAccessionSearchCache}

Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
