% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getAccessionSearchConfig}
\alias{getAccessionSearchConfig}
\title{Get the Accession Search Configuration}
\usage{
getAccessionSearchConfig()
}
\value{
a nested list containing the default accession search properties
}
\description{
Get the default configuration properties for the accession search 
function.  The default properties can be overriden by setting the 
global \code{breedbase.accession_search_config} option (Example: 
\code{options("breedbase.accession_search_config" = config)}).
}
\seealso{
Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
