% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{getAccessionSearchDBs}
\alias{getAccessionSearchDBs}
\title{Get Accession Search Databases}
\usage{
getAccessionSearchDBs()
}
\value{
Vector of supported Accession Search Databases, where each Database is 
a list with the following names:
\describe{
  \item{name}{Database name}
  \item{address}{Database BrAPI URL}
  \item{version}{Database BrAPI version}
  \item{call_limit}{Max number of concurrent connections to the Databse}
}
}
\description{
Get all of the supported BrAPI-compliant databases configured by the accession search server.
}
\details{
This function returns a vector of Accession Search Database objects, one of which can be used 
by the other accession search functions. The accession search server can be changed by setting 
the global \code{breedbase.accession_search_server} option (Example: 
\code{options("breedbase.accession_search_server" = "https://search.example.org")}).
}
\seealso{
\link{getAccessionSearchServer}

Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
