% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getAccessionSearchServer}
\alias{getAccessionSearchServer}
\title{Get the Accession Search Server}
\usage{
getAccessionSearchServer()
}
\value{
The URL to the Acccession Search server
}
\description{
Get the set URL for the BrAPI Accession Search server.  This value 
can be overriden by setting the global \code{breedbase.accession_search_server}
option (Example: \code{options("breedbase.accession_search_server" = "https://search.example.org")}).
}
\seealso{
Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
