% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getCountryCodes}
\alias{getCountryCodes}
\title{Get the Country Codes}
\usage{
getCountryCodes()
}
\value{
A vector of country codes
}
\description{
Get a vector of supported Location country codes.  This value can 
be overridden by setting the global \code{breedbase.country_codes} 
option (Example: \code{options("breedbase.country_codes" = country_codes)}).
}
\seealso{
Other Location: 
\code{\link{Location}},
\code{\link{Location-class}},
\code{\link{buildLocationTemplate}()},
\code{\link{geocodeLocation}()},
\code{\link{getLocationTypes}()},
\code{\link{lookupNOAAStationID}()},
\code{\link{writeLocationTemplate}()}
}
\concept{Location}
