% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getLocationTypes}
\alias{getLocationTypes}
\title{Get Location Types}
\usage{
getLocationTypes()
}
\value{
A vector of location types
}
\description{
Get a vector of supported Location types.  This value can be 
overridden by setting the global \code{breedbase.location_types} 
option (Example: \code{options("breedbase.location_types" = location_types)}).
}
\seealso{
Other Location: 
\code{\link{Location}},
\code{\link{Location-class}},
\code{\link{buildLocationTemplate}()},
\code{\link{geocodeLocation}()},
\code{\link{getCountryCodes}()},
\code{\link{lookupNOAAStationID}()},
\code{\link{writeLocationTemplate}()}
}
\concept{Location}
