% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{getSubmissionDirectories}
\alias{getSubmissionDirectories}
\title{Get Submission Directories}
\usage{
getSubmissionDirectories(dir, useMostRecent = TRUE)
}
\arguments{
\item{dir}{The path to the parent directory that contains the submission directores}

\item{useMostRecent}{When set to TRUE, only include the most recent submission directory for each trial}
}
\value{
vector of submission directories to parse
}
\description{
Get the list of submission directories from the parent directory
}
