% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getTrialDesignTypes}
\alias{getTrialDesignTypes}
\title{Get Trial Design Types}
\usage{
getTrialDesignTypes()
}
\value{
A vector of Trial design types
}
\description{
Get a vector of supported Trial design types.  This value can be 
overridden by setting the global \code{breedbase.design_types} 
option (Example: \code{options("breedbase.design_types" = design_types)}).
}
\seealso{
Other Trial: 
\code{\link{Plot-class}},
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial-class}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
