% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{getTrialTypes}
\alias{getTrialTypes}
\title{Get Trial Types}
\usage{
getTrialTypes()
}
\value{
A vector of Trial types
}
\description{
Get a vector of supported Trial types.  This value can be 
overridden by setting the global \code{breedbase.trial_types} 
option (Example: \code{options("breedbase.trial_types" = trial_types)}).
}
\seealso{
Other Trial: 
\code{\link{Plot}},
\code{\link{Plot-class}},
\code{\link{PlotData}},
\code{\link{PlotData-class}},
\code{\link{Trial}},
\code{\link{Trial-class}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
