% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\docType{class}
\name{Location-class}
\alias{Location-class}
\title{Location Class}
\description{
An S4 Class to represent a breeDBase Location
}
\section{Slots}{

\describe{
\item{\code{name}}{Unique location name (must not conflict with an existing location name)}

\item{\code{abbreviation}}{Location abbreviation}

\item{\code{country_code}}{ISO Alpha-3 country code}

\item{\code{country_name}}{Uppercase english short name from the ISO standard}

\item{\code{program}}{Name of one or more existing Breeding Programs}

\item{\code{type}}{Location type (Farm, Field, Greenhouse, Screenhouse, Lab, Storage, Other)}

\item{\code{latitude}}{Location latitude (decimal degrees)}

\item{\code{longitude}}{Location longitude (decimal degrees)}

\item{\code{altitude}}{Location elevation (meters)}

\item{\code{noaa_station_id}}{The NOAA Station ID that is nearest and most relevant to the location (ex: GHCND:US1NYTM0042)}
}}

