% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{Location}
\alias{Location}
\title{Create Location}
\usage{
Location(name = NULL, abbreviation = NULL, country_code = NULL,
  country_name = NULL, program = NULL, type = NULL,
  latitude = NULL, longitude = NULL, altitude = NULL)
}
\arguments{
\item{name}{Unique location name (must not conflict with an existing location name)}

\item{abbreviation}{Location abbreviation}

\item{country_code}{ISO Alpha-3 country code}

\item{country_name}{Uppercase english short name from the ISO standard}

\item{program}{Name of one or more existing Breeding Programs}

\item{type}{Location type (Farm, Field, Greenhouse, Screenhouse, Lab, Storage, Other)}

\item{latitude}{(optional) Location latitude (decimal degrees)}

\item{longitude}{(optional) Location longitude (decimal degrees)}

\item{altitude}{(optional) Location elevation (meters)}
}
\value{
Location
}
\description{
Create a Location containing all of the provided properties. 
Latitude, Longitude, and Altitude are optional and if not provided 
will be queried from the DataScienceToolkit API using the Location name.
}
\details{
The \code{\link{geocodeLocation}} function can be used to obtain lat/lon/alt 
properties from a more specific location name (such as a street address), the 
results of which can be used to create the Location.
}
\examples{
# Use the Location name to query lat/lon/alt properties
location <- Location(
     "Batavia, NY",
     "BAT",
     "USA",
     "United States of America",
     "Cornell",
     "Field"
)

# Geocode an address to use as lat/lon/alt properties
geo <- geocodeLocation("2 Caldwell Drive, Ithaca, NY")
location <- Location(
     "Caldwell - Ithaca, NY", 
     "ITH_CALD",
     "USA",
     "United States of America",
     "Cornell",
     "Field",
     geo$latitude,
     geo$longitude,
     geo$altitude
)

}
