% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{lookupNOAAStationID}
\alias{lookupNOAAStationID}
\title{Lookup NOAA Station ID}
\usage{
lookupNOAAStationID(lat, lon, radius = 10)
}
\arguments{
\item{lat}{Location latitiude (decimal degrees)}

\item{lon}{Location longitude (decimal degrees)}

\item{radius}{Search radius (miles, default=10)}
}
\description{
Lookup the NOAA Station ID for the specified location.  The 
Location must be specified by its latitude and longitude 
(as decimal degrees) - if these are unknown the \code{\link{geocodeLocation}} 
function can be used to get the coordinates of a location by 
address, town, etc.
}
\details{
The query will search for NOAA stations within the specified radius 
centered around the location position.  If multiple stations are found,
the ID of the first/closest station is returned.
}
