% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{matchAccessions}
\alias{matchAccessions}
\title{Attempt to Find Best Match for Set of Accessions}
\usage{
matchAccessions(db, terms, config = getBBOption("accession_search_config"))
}
\description{
Use performAccessionSearch() to obtain a data frame of potential matches for a set of accessions, then identify the root accession name for all matches.
}
\details{
This is intended for use cases such as identifying the breedbase-stored names of a trial given a list of lines in that trial from a collaborator or other source.

The accession search can be configured by providing a list of accession search properties.
}
