% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeData}
\alias{mergeData}
\title{Merge Data}
\usage{
mergeData(dir, filename, key, useMostRecent, keepUniqueDuplicates)
}
\arguments{
\item{dir}{path to parent directory of the submissions}

\item{filename}{name of the file to merge}

\item{key}{the name of the key column for finding duplicates}

\item{useMostRecent}{When set to TRUE, only use the most recent submission directory for each trial}

\item{keepUniqueDuplicates}{When set to TRUE, the script will keep all unique versions of duplicate rows (and won't prompt the user to pick)}
}
\description{
Perform the actual merging of the data with the provided properties
}
