% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{mergeLocations}
\alias{mergeLocations}
\title{Merge Locations}
\usage{
mergeLocations(dir = ".", useMostRecent = TRUE, keepUniqueDuplicates = FALSE)
}
\arguments{
\item{dir}{The path to the parent directory that contains the submission directories (default=current directory)}

\item{useMostRecent}{When set to TRUE, only use the most recent submission directory for each trial}

\item{keepUniqueDuplicates}{When set to TRUE, the script will keep all unique versions of duplicate rows (and won't prompt the user to pick)}
}
\description{
Create a merged location template from multiple submissions
}
