% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\docType{class}
\name{Pedigree-class}
\alias{Pedigree-class}
\title{Pedigree Class}
\description{
An S4 Class to represent a breeDBase pedigree
}
\section{Slots}{

\describe{
\item{\code{progeny_name}}{Name or synonym of the Accession to add pedigree information to (must exist in database)}

\item{\code{female_parent_accession}}{Name of female parent Accession}

\item{\code{male_parent_accession}}{Name of male parent Accession}

\item{\code{type}}{Cross type (biparental, self, open, sib)}
}}

\seealso{
Other Accession: 
\code{\link{Accession}},
\code{\link{Accession-class}},
\code{\link{Pedigree}},
\code{\link{buildAccessionTemplate}()},
\code{\link{buildPedigreeTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writeAccessionTemplate}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
