% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{Pedigree}
\alias{Pedigree}
\title{Create Pedigree}
\usage{
Pedigree(
  progeny = NULL,
  female_parent = NULL,
  male_parent = NULL,
  type = "biparental"
)
}
\arguments{
\item{progeny}{The progeny Accession or Accession name (or synonym)}

\item{female_parent}{The female parent Accession or Accession name}

\item{male_parent}{The male parent Accession or Accession name}

\item{type}{(optional, default = 'biparental') The cross type (biparental, self, open, sib)}
}
\value{
Pedigree
}
\description{
Create a Pedigree for the Accession specified as `progeny` 
with the given female and male parents.
}
\examples{
# Create a Pedigree from names
pedigree1 <- Pedigree('SL18-UCONN-S1', 'SA18-CB-S10-FG1', 'SL18-SF-S19-MG2')

# Create a Pedigree from existing Accessions
offspring <- Accession('SL18-UCONN-S1', 'Saccharina latissima')
mother <- Accession('SA18-CB-S10-FG1', 'Saccharina angustissima')
father <- Accession('SL18-SF-S19-MG2', 'Saccharina latissima')
pedigree2 <- Pedigree(offspring, mother, father)

}
\seealso{
Other Accession: 
\code{\link{Accession}},
\code{\link{Accession-class}},
\code{\link{Pedigree-class}},
\code{\link{buildAccessionTemplate}()},
\code{\link{buildPedigreeTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writeAccessionTemplate}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
