% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\docType{class}
\name{Plot-class}
\alias{Plot-class}
\title{Plot Class}
\description{
An S4 Class to represent a breeDBase phenotyping trial plot
}
\details{
The `plot_name`, `accession_name`, `plot_number`, and `block_number` 
are required, all other fields are optional
}
\section{Slots}{

\describe{
\item{\code{plot_name}}{The unique name for the plot (must be unique across entire database. this is often a concatenation of the trial name, the accession name, and the plot number.)}

\item{\code{accession_name}}{The name of the accession being tested in the plot (must exist in the database)}

\item{\code{plot_number}}{The sequential number for the plot in the field (e.g. 1001, 1002, 2001, 2002). these numbers should be unique for the trial.}

\item{\code{block_number}}{A design parameter indicating which block the plot is in}

\item{\code{is_a_control}}{(optional) TRUE if the plot is a control.  Generally, you will have accessions that are controls, so you should indicate the plots that that accession is in as a control.}

\item{\code{rep_number}}{(optional) replicate number}

\item{\code{range_number}}{(optional) range number, often synonymous with col_number}

\item{\code{row_number}}{(optional) row number. if the field is a grid, this represents the y coordinate, required for field map generation. the top left plot should be row 1, column 1}

\item{\code{col_number}}{(optional) column number. if the field is a grid, this represents the x coordinate. sometimes called range_number, required for field map generation. the top left plot shuold be row 1, column 1}

\item{\code{seedlot_name}}{(optional) the seedlot from where the planted seed originated. must exist in the database}

\item{\code{num_seed_per_plot}}{(optional) number seeds per plot. seed is transferred from seedlot mentioned in seedlot_name.}

\item{\code{weight_gram_seed_per_plot}}{(optional) weight in gram of seeds in plot. seed is transferred from seedlot mentioned in seedlot name.}

\item{\code{treatments}}{(optional) A named list where the name is the name of a treatment (e.g. inoculated, drought, etc) and the value should be TRUE if the treatment was applied to the plot or FALSE if not.}
}}

\seealso{
Other Trial: 
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial-class}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
