% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdatas.R
\name{PlotData}
\alias{PlotData}
\title{Create PlotData}
\usage{
PlotData(plot_name = NULL, observations = list(), notes = NA_character_)
}
\arguments{
\item{plot_name}{The unique name for the plot (must already exist as a Plot in a Trial)}

\item{observations}{List of observations where the key is the trait name and the value is the observed trait value}

\item{notes}{(optional) Additional plot notes}
}
\value{
PlotData
}
\description{
Create a PlotData object that contains phenotype observations 
for a single plot in a phenotyping trial.  The Plot Name and 
Observations are required.
}
\details{
The Observations take the form of a named list where the list item name / key is the 
full trait name, including the trait ID (ex: "Blade length cm|CO_360:0000240") and the 
list item value is the observed trait value (in the correct units).
}
\examples{
plotdata1 <- PlotData("FARM-2019-UNH_PLOT1", list("Blade length cm|CO_360:0000240" = 24, "Blade width cm|CO_360:0000241" = 5))
plotdata2 <- PlotData("FARM-2019-UNH_PLOT2", list("Blade length cm|CO_360:0000240" = 27, "Blade width cm|CO_360:0000241" = 4))

}
