% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Progeny.R
\docType{class}
\name{Progeny-class}
\alias{Progeny-class}
\title{Progeny Class}
\description{
An S4 Class to represent a breeDBase progeny / pedigree
}
\section{Slots}{

\describe{
\item{\code{progeny_name}}{Name or synonym of the Accession to add pedigree information to (must exist in database)}

\item{\code{female_parent_accession}}{Name of female parent Accession}

\item{\code{male_parent_accession}}{Name of male parent Accession}

\item{\code{type}}{Cross type (biparental, self, open, sib)}
}}

