% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenies.R
\name{Progeny}
\alias{Progeny}
\title{Create Progeny}
\usage{
Progeny(progeny = NULL, female_parent = NULL, male_parent = NULL,
  type = "biparental")
}
\arguments{
\item{progeny}{The progeny Accession or Accession name (or synonym)}

\item{female_parent}{The female parent Accession or Accession name}

\item{male_parent}{The male parent Accession or Accession name}

\item{type}{(optional, default = 'biparental') The cross type (biparental, self, open, sib)}
}
\value{
Progeny
}
\description{
Create a Progeny / Pedigree for the Accession specified 
as `progeny` with the given female and male parents.
}
\examples{
# Create a Progeny from names
progeny1 <- Progeny('SL18-UCONN-S1', 'SA18-CB-S10-FG1', 'SL18-SF-S19-MG2')

# Create a Progeny from existing Accessions
offspring <- Accession('SL18-UCONN-S1', 'Saccharina latissima')
mother <- Accession('SA18-CB-S10-FG1', 'Saccharina angustissima')
father <- Accession('SL18-SF-S19-MG2', 'Saccharina latissima')
progeny2 <- Progeny(offspring, mother, father)

}
