% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove Duplicates}
\usage{
removeDuplicates(merged, key_column, keepUniqueDuplicates)
}
\arguments{
\item{merged}{The merged tibble that contains all of the rows}

\item{key_column}{The name of the column that will be used to find unique values}

\item{keepUniqueDuplicates}{When TRUE, the script will not prompt to pick a unique duplicate row, but will keep all of them}
}
\value{
a tibble with the duplicate rows removed
}
\description{
Remove duplicate rows that have the same value in the specified key column
}
