% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{setTrialPlots}
\alias{setTrialPlots}
\title{Set Trial Plots}
\usage{
setTrialPlots(trial = NULL, plots = NULL)
}
\arguments{
\item{trial}{A Trial to add the plots to}

\item{plots}{A vector of Plots to add to the Trial}
}
\value{
Trial
}
\description{
Set the provided vector of Plots as the `plots` property 
of the provided Trial
}
