% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{updateAccessionSearchCache}
\alias{updateAccessionSearchCache}
\title{Update Accession Search Cache}
\usage{
updateAccessionSearchCache(db)
}
\arguments{
\item{db}{Accession Search Database}
}
\value{
nothing
}
\description{
Update the cache of database terms for the provided Accession Search Database.
}
\seealso{
Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()}
}
\concept{accessionSearch}
