% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessions.R
\name{writeAccessionTemplate}
\alias{writeAccessionTemplate}
\title{Write Accession Template}
\usage{
writeAccessionTemplate(input = NULL, output = NULL, chunk = NULL)
}
\arguments{
\item{input}{Either a vector of Accessions to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .xls file}

\item{chunk}{Chunk the file into parts with up to `chunk` number of lines per file}
}
\description{
Write a breeDBase upload template (.xls file) for accessions
}
\seealso{
Other Accession: 
\code{\link{Accession-class}},
\code{\link{Accession}},
\code{\link{Pedigree-class}},
\code{\link{Pedigree}},
\code{\link{buildAccessionTemplate}()},
\code{\link{buildPedigreeTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
