% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{writeLocationTemplate}
\alias{writeLocationTemplate}
\title{Write Location Template}
\usage{
writeLocationTemplate(input = NULL, output = NULL, chunk = NULL)
}
\arguments{
\item{input}{Either a vector of Locations to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .xls file}

\item{chunk}{Chunk the file into parts with up to `chunk` number of lines per file}
}
\description{
Write a breeDBase upload template (.xls file) for locations
}
