% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{writePedigreeTemplate}
\alias{writePedigreeTemplate}
\title{Write Pedigree Template}
\usage{
writePedigreeTemplate(input = NULL, output = NULL, chunk = NULL)
}
\arguments{
\item{input}{Either a vector of Pedigrees to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .txt file}

\item{chunk}{Chunk the file into parts with up to `chunk` number of lines per file}
}
\description{
Write a breeDBase upload template (.txt tab delimited file) for pedigrees
}
