% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdatas.R
\name{writePlotDataTemplate}
\alias{writePlotDataTemplate}
\title{Write PlotData Template}
\usage{
writePlotDataTemplate(input = NULL, output = NULL)
}
\arguments{
\item{input}{Either a vector of PlotDatas to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .xls file}
}
\description{
Write a breeDBase upload template (.xls file) for plot data
}
