% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{writePlotTemplate}
\alias{writePlotTemplate}
\title{Write Plot Template}
\usage{
writePlotTemplate(input = NULL, output = NULL)
}
\arguments{
\item{input}{Either a vector of Plots to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .xls file}
}
\description{
Write a breeDBase upload template (.xls file) for plots
}
\seealso{
Other Trial: 
\code{\link{Plot-class}},
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial-class}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
