% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessions.R
\name{Accession}
\alias{Accession}
\title{Create Accession}
\usage{
Accession(accession_name = NULL, species_name = NULL, properties = list())
}
\arguments{
\item{accession_name}{Accession Name}

\item{species_name}{Species Name}

\item{properties}{(optional) Additional accession properties (as a named list)

See Class \linkS4class{Accession} for all optional accession properties}
}
\value{
Accession
}
\description{
Create an Accession containing all of the provided properties. 
Accession Name and Species Name are required.  All optional 
properties can be added as a named value in the properties list.
}
\examples{
accession <- Accession(
     "JERRY", 
     "Triticum aestivum",
     list(
         synonyms = c("ND9257", "PI632433"),
         institute_codes = "NDSU",
         organization_names = "North Dakota State University"
     )
)

}
