% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigrees.R
\name{buildPedigreeTemplate}
\alias{buildPedigreeTemplate}
\title{Build Pedigree Template}
\usage{
buildPedigreeTemplate(pedigrees = NULL)
}
\arguments{
\item{pedigrees}{Vector of Pedigrees to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBase upload 
template for the provided pedigrees
}
\seealso{
Other Accession: 
\code{\link{Accession}},
\code{\link{Accession-class}},
\code{\link{Pedigree}},
\code{\link{Pedigree-class}},
\code{\link{buildAccessionTemplate}()},
\code{\link{getSupportedAccessionProperties}()},
\code{\link{writeAccessionTemplate}()},
\code{\link{writePedigreeTemplate}()}
}
\concept{Accession}
