% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdatas.R
\name{buildPlotDataTemplate}
\alias{buildPlotDataTemplate}
\title{Build PlotData Template}
\usage{
buildPlotDataTemplate(plotdata = NULL)
}
\arguments{
\item{plotdata}{Vector of PlotData to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBasae upload
template for the provided plotdata.  This will be a 'Simple' 
spreadsheet format with data recorded at the plot-level.
}
\seealso{
Other Trial: 
\code{\link{Plot}},
\code{\link{Plot-class}},
\code{\link{PlotData}},
\code{\link{PlotData-class}},
\code{\link{Trial}},
\code{\link{Trial-class}},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
