% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{buildPlotTemplate}
\alias{buildPlotTemplate}
\title{Build Plot Template}
\usage{
buildPlotTemplate(plots = NULL)
}
\arguments{
\item{plots}{Vector of Plots to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBase upload 
template for the provided plots
}
\seealso{
Other Trial: 
\code{\link{Plot}},
\code{\link{Plot-class}},
\code{\link{PlotData}},
\code{\link{PlotData-class}},
\code{\link{Trial}},
\code{\link{Trial-class}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
