% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{buildTrialTemplate}
\alias{buildTrialTemplate}
\title{Build Trial Template}
\usage{
buildTrialTemplate(trials = NULL)
}
\arguments{
\item{trials}{Vector of Trials to add to the template}
}
\value{
A \code{tibble} representation of the upload template
}
\description{
Create a \code{tibble} representing the breeDBase upload 
template for the provided trials.  The Trial must have 
plots set in order to add it to the template.
}
\seealso{
Other Trial: 
\code{\link{Plot-class}},
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial-class}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
