% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_buac_kgHa}
\alias{convert_buac_kgHa}
\title{Convert bu/ac to kg/Ha}
\usage{
convert_buac_kgHa(x, crop = NULL)
}
\arguments{
\item{x}{grain yield in bu/ac}

\item{crop}{crop type ("barley", "oat", "wheat")}
}
\value{
grain yield in kg/Ha
}
\description{
Convert grain yield units from bu/ac to kg/Ha
}
\seealso{
Other conversions: 
\code{\link{convert_buac_gm2}()},
\code{\link{convert_gm2_kgHa}()},
\code{\link{convert_in_cm}()},
\code{\link{convert_kgHa_gm2}()},
\code{\link{convert_kghl_gL}()},
\code{\link{convert_lbsbu_gL}()},
\code{\link{convert_scale}()},
\code{\link{convert_sqft_m2}()}
}
\concept{conversions}
