% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_gm2_kgHa}
\alias{convert_gm2_kgHa}
\title{Convert g/m2 to kg/Ha}
\usage{
convert_gm2_kgHa(x)
}
\arguments{
\item{x}{grain yield in g/m2}
}
\value{
grain yield in kg/Ha
}
\description{
Convert grain yield units from g/m2 to kg/Ha
}
