% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_in_cm}
\alias{convert_in_cm}
\title{Convert in to cm}
\usage{
convert_in_cm(x)
}
\arguments{
\item{x}{height in inches}
}
\value{
height in cm
}
\description{
Convert height / length units from in to cm
}
