% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_kgHa_gm2}
\alias{convert_kgHa_gm2}
\title{Convert kg/Ha to g/m2}
\usage{
convert_kgHa_gm2(x)
}
\arguments{
\item{x}{grain yield in kg/Ha}
}
\value{
grain yield in g/m2
}
\description{
Convert grain yield units from kg/Ha to g/m2
}
