% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_kghl_gL}
\alias{convert_kghl_gL}
\title{Convert kg/hl to g/l}
\usage{
convert_kghl_gL(x)
}
\arguments{
\item{x}{test weight in kg/hl}
}
\value{
test weight in g/L
}
\description{
Convert test weight units from kg/hl to g/L
}
