% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_lbsbu_gL}
\alias{convert_lbsbu_gL}
\title{Convert lbs/bu to g/L}
\usage{
convert_lbsbu_gL(x)
}
\arguments{
\item{x}{test weight in lbs/bu}
}
\value{
test weight in g/L
}
\description{
Convert test weight units from lbs/bu to g/L
}
\seealso{
Other conversions: 
\code{\link{convert_buac_gm2}()},
\code{\link{convert_buac_kgHa}()},
\code{\link{convert_gm2_kgHa}()},
\code{\link{convert_in_cm}()},
\code{\link{convert_kgHa_gm2}()},
\code{\link{convert_kghl_gL}()},
\code{\link{convert_scale}()},
\code{\link{convert_sqft_m2}()}
}
\concept{conversions}
