% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_sqft_m2}
\alias{convert_sqft_m2}
\title{Convert x/sq ft to x/m2}
\usage{
convert_sqft_m2(x)
}
\arguments{
\item{x}{rate per square foot}
}
\value{
rate per square meter
}
\description{
Convert a rate per square foot to a 
rate per square meters
}
