% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{getAccessionSearchDBs}
\alias{getAccessionSearchDBs}
\title{Get Accession Search Database Names}
\usage{
getAccessionSearchDBs()
}
\value{
Vector of supported database names
}
\description{
Get the names of all of the supported BrAPI-compliant databases configured 
by the accession search server.
}
\details{
Using the database name, the \code{\link{getAccessionSearchDB}} function 
can be used to get the Accession Search Database object needed by the other accession 
search functions. The accession search server can be changed by setting the global 
\code{breedbase.accession_search_server} option (Example: 
\code{options("breedbase.accession_search_server" = "https://search.example.org")}).
}
\seealso{
\link{getAccessionSearchDB}

\link{getAccessionSearchServer}

Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{performAccessionSearch}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
