% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessionSearch.R
\name{performAccessionSearch}
\alias{performAccessionSearch}
\title{Perform an Accession Search}
\usage{
performAccessionSearch(
  db,
  terms,
  config = getBBOption("accession_search_config")
)
}
\arguments{
\item{db}{Accession Search Database}

\item{terms}{Vector of accession search terms}

\item{config}{Accession Search Configuration}
}
\value{
a \code{tibble} containing the search results. Example:\tabular{llllll}{
   \strong{search_term} \tab \strong{search_routine} \tab \strong{germplasm_name} \tab \strong{germplasm_id} \tab \strong{database_term} \tab \strong{database_term_type} \cr
   jerry \tab Exact Match \tab JERRY \tab 230227 \tab JERRY \tab name \cr
   SY-Gold \tab Exact Match \tab 00X0100-51 \tab 232633 \tab Sy-Gold \tab synonym \cr
   PU0128A1_36 \tab Remove Punctuation \tab 0218A1-36 \tab 238801 \tab PU0128A1-36 \tab synonym \cr
}
}
\description{
Search the provided Accession Search Database to find database terms
that match the provided search terms.
}
\details{
The accession search can be configured by providing a set of accession search
configuration properties, which is a nested list with the following names:\tabular{lll}{
   \strong{NAME} \tab \strong{DEFINITION} \tab \strong{DEFAULT VALUE} \cr
   \strong{database_terms} \tab a list with the names: name, synonyms, accession_numbers \tab  \cr
   database_terms$name \tab flag to include accession names in the accession search \tab TRUE \cr
   database_terms$synonyms \tab flag to include accession synonyms in the accession search \tab TRUE \cr
   database_terms$accession_numbers \tab flag to include the accession numbers in the accession search \tab TRUE \cr
   \strong{search_routines} \tab a list with the names: exact, punctuation, substring, prefix, edit_distance \tab  \cr
   search_routines$exact \tab flag to use the exact name search routine in the accession search \tab TRUE \cr
   search_routines$punctuation \tab flag to use the punctuation search routine in the accession search \tab TRUE \cr
   search_routines$substring \tab flag to use the substring search routine in the accession search \tab TRUE \cr
   search_routines$prefix \tab flag to use the prefix search routine in the accession search \tab FALSE \cr
   search_routines$edit_distance \tab flag to use the edit distance search routine in the accession search \tab FALSE \cr
   \strong{search_routine_options} \tab a list that has names of the supported search routines with values set as a list of the search routine options \tab  \cr
   search_routine_options$substring$substring_length_min \tab the minimum length of a term to be included as a substring match \tab 3 \cr
   search_routine_options$prefix$prefixes \tab default prefixes to include \tab [] \cr
   search_routine_options$prefix$find_db_prefixes \tab scan the database terms to find common prefixes \tab TRUE \cr
   search_routine_options$prefix$prefix_length_min \tab when finding db prefixes, the minimum length of a prefix to include \tab 2 \cr
   search_routine_options$prefix$prefix_length_max \tab when finding db prefixes, the maximum length of a prefix to include \tab 5 \cr
   search_routine_options$prefix$threshold \tab when finding db prefixes, the minimum number of times a prefix is used before it is included \tab 250 \cr
   search_routine_options$edit_distance$max_edit_distance \tab the maximum number of changes for the edit distance comparison \tab 2 \cr
   \strong{case_sensitive} \tab flag to perform a case-sensitive search \tab TRUE \cr
   \strong{return_records} \tab flag to include the germplasm records with the search results \tab FALSE \cr
}
}
\seealso{
Other accessionSearch: 
\code{\link{createAccessionSearchDB}()},
\code{\link{getAccessionSearchCacheStatus}()},
\code{\link{getAccessionSearchConfig}()},
\code{\link{getAccessionSearchDBs}()},
\code{\link{getAccessionSearchDB}()},
\code{\link{getAccessionSearchServer}()},
\code{\link{updateAccessionSearchCache}()}
}
\concept{accessionSearch}
