% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{setTrialPlots}
\alias{setTrialPlots}
\title{Set Trial Plots}
\usage{
setTrialPlots(trial = NULL, plots = NULL)
}
\arguments{
\item{trial}{A Trial to add the plots to}

\item{plots}{A vector of Plots to add to the Trial}
}
\value{
Trial
}
\description{
Set the provided vector of Plots as the `plots` property 
of the provided Trial
}
\seealso{
Other Trial: 
\code{\link{Plot-class}},
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial-class}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
