% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trial.R
\docType{class}
\name{Trial-class}
\alias{Trial-class}
\title{Trial Class}
\description{
An S4 Class to represent a breeDBase phenotyping trial
}
\details{
The `trial_name`, `breeding_program`, `location`, `year`, `design_type`, and 
`description` are all required, all other fields are optional.
}
\section{Slots}{

\describe{
\item{\code{trial_name}}{A name for the trial (must be unique across the entire database)}

\item{\code{breeding_program}}{The name of the breeding program that managed the trial (must exist in database)}

\item{\code{location}}{The name of the location where the trial was held (must exist in database)}

\item{\code{year}}{The year the trial was held}

\item{\code{design_type}}{The shorthand for the design type (CRD, RCBD, Alpha, Augmented, MAD, Westcott, Lattice)}

\item{\code{description}}{Additional text with any other relevant information about the trial}

\item{\code{trial_type}}{(optional) The name of the trial type (must exist in the database)}

\item{\code{plot_width}}{(optional) The plot width in meters}

\item{\code{plot_length}}{(optional) The plot length in meters}

\item{\code{field_size}}{(optional) The field size in hectares}

\item{\code{planting_date}}{(optional) The date of planting in YYYY-MM-DD format}

\item{\code{harvest_date}}{(optional) The date of harvest in YYYY-MM-DD format}

\item{\code{plots}}{(optional) A vector of Plots that are used in the trial}
}}

\seealso{
Other Trial: 
\code{\link{Plot-class}},
\code{\link{PlotData-class}},
\code{\link{PlotData}},
\code{\link{Plot}},
\code{\link{Trial}},
\code{\link{buildPlotDataTemplate}()},
\code{\link{buildPlotTemplate}()},
\code{\link{buildTrialTemplate}()},
\code{\link{createPlots}()},
\code{\link{getTrialDesignTypes}()},
\code{\link{getTrialTypes}()},
\code{\link{printPlots}()},
\code{\link{setTrialPlots}()},
\code{\link{writePlotDataTemplate}()},
\code{\link{writePlotTemplate}()},
\code{\link{writeTrialTemplate}()}
}
\concept{Trial}
