% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessions.R
\name{writeAccessionTemplate}
\alias{writeAccessionTemplate}
\title{Write Accession Template}
\usage{
writeAccessionTemplate(input = NULL, output = NULL)
}
\arguments{
\item{input}{Either a vector of Accessions to include in the template OR 
a \code{tibble} representation of the upload template}

\item{output}{The file path to the output .xls file}
}
\description{
Write a breeDBase upload template (.xls file) for accessions
}
